changed tablesorter directive to hometablesorter, added displayShow tablesorter directive, changed episodes in displayShow JSON to seasons object with array of seasons
