+++ gui/slick/templates/home.html
@@ -1,48 +1,47 @@
 <table style="width: 100%;" class="home-header">
     <tr>
         <td align="right">
-                <div ng-if="layout !== 'poster'">
-                    <span class="show-option">
-                        <button id="popover" type="button" class="btn btn-inline">Select Columns <b class="caret"></b></button>
-                    </span>
+            <div ng-if="layout !== 'poster'">
+                <span class="show-option">
+                    <button tablesorterpopover type="button" class="btn btn-inline">Select Columns <b class="caret"></b></button>
+                </span>
 
-                    <span class="show-option">
-                        <button type="button" class="resetsorting btn btn-inline">Clear Filter(s)</button>
-                    </span>
-                </div>
+                <span class="show-option">
+                    <button type="button" class="resetsorting btn btn-inline">Clear Filter(s)</button>
+                </span>
+            </div>
 
-                <div ng-if="layout === 'poster'">
-                    <span class="show-option"> Poster Size:
-                        <div style="width: 100px; display: inline-block; margin-left: 7px;" id="posterSizeSlider"></div>
-                    </span>
+            <div ng-if="layout === 'poster'">
+                <span class="show-option"> Poster Size:
+                    <div style="width: 100px; display: inline-block; margin-left: 7px;" id="posterSizeSlider"></div>
+                </span>
 
-                    <span class="show-option">
-                        <input id="filterShowName" class="form-control form-control-inline input-sm input200" type="search" placeholder="Filter Show Name">
-                    </span>
+                <span class="show-option">
+                    <input id="filterShowName" class="form-control form-control-inline input-sm input200" type="search" placeholder="Filter Show Name">
+                </span>
 
-                    <span class="show-option"> Sort By:
-                        <select id="postersort" class="form-control form-control-inline input-sm">
-                            <option value="name" data-sort="${srRoot}/setPosterSortBy/?sort=name" ${('', 'selected="selected"')[sickbeard.POSTER_SORTBY == 'name']}>Name</option>
-                            <option value="date" data-sort="${srRoot}/setPosterSortBy/?sort=date" ${('', 'selected="selected"')[sickbeard.POSTER_SORTBY == 'date']}>Next Episode</option>
-                            <option value="network" data-sort="${srRoot}/setPosterSortBy/?sort=network" ${('', 'selected="selected"')[sickbeard.POSTER_SORTBY == 'network']}>Network</option>
-                            <option value="progress" data-sort="${srRoot}/setPosterSortBy/?sort=progress" ${('', 'selected="selected"')[sickbeard.POSTER_SORTBY == 'progress']}>Progress</option>
-                        </select>
-                    </span>
+                <span class="show-option"> Sort By:
+                    <select id="postersort" class="form-control form-control-inline input-sm">
+                        <option value="name" data-sort="${srRoot}/setPosterSortBy/?sort=name" ${('', 'selected="selected"')[sickbeard.POSTER_SORTBY == 'name']}>Name</option>
+                        <option value="date" data-sort="${srRoot}/setPosterSortBy/?sort=date" ${('', 'selected="selected"')[sickbeard.POSTER_SORTBY == 'date']}>Next Episode</option>
+                        <option value="network" data-sort="${srRoot}/setPosterSortBy/?sort=network" ${('', 'selected="selected"')[sickbeard.POSTER_SORTBY == 'network']}>Network</option>
+                        <option value="progress" data-sort="${srRoot}/setPosterSortBy/?sort=progress" ${('', 'selected="selected"')[sickbeard.POSTER_SORTBY == 'progress']}>Progress</option>
+                    </select>
+                </span>
 
-                    <span class="show-option"> Direction:
-                        <select id="postersortdirection" class="form-control form-control-inline input-sm">
-                            <option value="true" data-sort="${srRoot}/setPosterSortDir/?direction=1" ${('', 'selected="selected"')[sickbeard.POSTER_SORTDIR == 1]}>Ascending </option>
-                            <option value="false" data-sort="${srRoot}/setPosterSortDir/?direction=0" ${('', 'selected="selected"')[sickbeard.POSTER_SORTDIR == 0]}>Descending</option>
-                        </select>
-                    </span>
-                </div>
-
-                <span class="show-option"> Layout:
-                    <select ng-model="layout" ng-options="layout as layout for layout in layouts" name="layout" class="form-control form-control-inline input-sm">
-                        {{layout}}
+                <span class="show-option"> Direction:
+                    <select id="postersortdirection" class="form-control form-control-inline input-sm">
+                        <option value="true" data-sort="${srRoot}/setPosterSortDir/?direction=1" ${('', 'selected="selected"')[sickbeard.POSTER_SORTDIR == 1]}>Ascending </option>
+                        <option value="false" data-sort="${srRoot}/setPosterSortDir/?direction=0" ${('', 'selected="selected"')[sickbeard.POSTER_SORTDIR == 0]}>Descending</option>
                     </select>
                 </span>
             </div>
+
+            <span class="show-option"> Layout:
+                <select ng-model="layout" ng-options="layout as layout for layout in layouts" name="layout" class="form-control form-control-inline input-sm">
+                    {{layout}}
+                </select>
+            </span>
         </td>
     </tr>
 </table>
