+++ gui/slick/templates/partials/home/banner.html
@@ -1,4 +1,4 @@
-<table ng-repeat="showList in showLists" class="tablesorter" cellspacing="1" border="0" cellpadding="0">
+<table hometablesorter ng-repeat="showList in showLists" cellspacing="1" border="0" cellpadding="0">
     <thead>
         <tr>
             <th class="nowrap">Next Ep</th>
@@ -15,12 +15,12 @@
     <tbody>
         <tr ng-controller="bannerShowController" ng-repeat="(i, show) in showList.shows" ng-class-odd="'odd'" ng-class-even="'even'">
             <td ng-if="show.airsNext" align="center" class="nowrap">
-                <time datetime="{{show.airsNext}}" class="date">{{show.airsNext}}</time>
+                <time timeago datetime="{{show.airsNext}}" class="date">{{show.airsNext}}</time>
             </td>
             <td ng-if="!show.airsNext" align="center" class="nowrap"></td>
 
             <td ng-if="show.airsPrev" align="center" class="nowrap">
-                <time datetime="{{show.airsPrev}}" class="date">{{show.airsPrev}}</time>
+                <time timeago datetime="{{show.airsPrev}}" class="date">{{show.airsPrev}}</time>
             </td>
             <td ng-if="!show.airsPrev" align="center" class="nowrap"></td>
 
@@ -28,16 +28,16 @@
                 <span style="display: none;">{{show.name}}</span>
                 <div class="imgbanner banner">
                     <a href="{{'#/displayShow?showId=' + show.indexerId}}">
-                    <img src="{{'/showPoster/?show=' + show.indexerId + '&amp;which=banner'}}" class="banner" alt="{{show.indexerId}}" title="{{show.name}}"/>
+                    <img ng-src="{{'/showPoster/?show=' + show.indexerId + '&amp;which=banner'}}" class="banner" alt="{{show.indexerId}}" title="{{show.name}}"/>
                 </div>
             </td>
 
             <td ng-if="show.network" align="center">
-                <span title="{{show.network}}" class="hidden-print"><img id="network" width="54" height="27" src="{{'/showPoster/?show=' + show.indexerId + '&amp;which=network'}}" alt="{{show.network}}" title="{{show.network}}" /></span>
+                <span title="{{show.network}}" class="hidden-print"><img id="network" width="54" height="27" ng-src="{{'/showPoster/?show=' + show.indexerId + '&amp;which=network'}}" alt="{{show.network}}" title="{{show.network}}" /></span>
                 <span class="visible-print-inline">${curShow.network}</span>
             </td>
             <td ng-if="!show.network" align="center">
-                <span title="No Network" class="hidden-print"><img id="network" width="54" height="27" src="/images/network/nonetwork.png" alt="No Network" title="No Network" /></span>
+                <span title="No Network" class="hidden-print"><img id="network" width="54" height="27" ng-src="/images/network/nonetwork.png" alt="No Network" title="No Network" /></span>
                 <span class="visible-print-inline">No Network</span>
             </td>
 
@@ -53,8 +53,8 @@
             <td align="center" data-show-size="{{show.size.raw}}">{{show.size.pretty}}</td>
 
             <td align="center">
-                <img ng-if="show.paused || show.status.status.indexOf('ended') >= 0" src="/images/no16.png" alt="Yes" width="16" height="16" />
-                <img ng-if="show.status.indexOf('Continuing') >= 0" src="/images/yes16.png" alt="No" width="16" height="16" />
+                <img ng-if="show.paused || show.status.status.indexOf('ended') >= 0" ng-src="/images/no16.png" alt="Yes" width="16" height="16" />
+                <img ng-if="show.status.indexOf('Continuing') >= 0" ng-src="/images/yes16.png" alt="No" width="16" height="16" />
             </td>
 
             <td align="center">{{show.status}}</td>
