+++ gui/slick/templates/partials/home/small.html
@@ -1,4 +1,4 @@
-<table ng-repeat="showList in showLists" class="tablesorter" cellspacing="1" border="0" cellpadding="0">
+<table hometablesorter ng-repeat="showList in showLists" class="tablesorter" cellspacing="1" border="0" cellpadding="0">
     <thead>
         <tr>
             <th class="nowrap">Next Ep</th>
